requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
openpyxl==3.1.5
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.5
XlsxWriter==3.2.0
waitress==3.0.0
pandas==2.3.3
Werkzeug==3.1.3
